Add travis.yml file to run continuous integration tests.
Install pytest for Travis tests.
Move python dependencies to requirements.txt.

Use mmh3 instead of pyhash

pyhash cannot be installed on Travis due to an outdated check
for libboost_python.so. mmh3 seems to be better maintained and
more widely used.

+++ install.sh
@@ -1,3 +1,6 @@
+#!/bin/bash
+set -ev
+
 echo "Would you like to install Adobe Flash Player? (Only required for crawls with Flash) [y,N]"
 read -s -n 1 response
 if [[ $response = "" ]] || [ $response == 'n' ] || [ $response == 'N' ]; then
@@ -14,24 +17,19 @@ fi
 
 sudo apt-get update
 
-sudo apt-get install firefox htop git python-dev libxml2-dev libxslt-dev libffi-dev libssl-dev build-essential xvfb libboost-python-dev libleveldb1 libleveldb-dev libjpeg-dev
+sudo apt-get install -y firefox htop git python-dev libxml2-dev libxslt-dev libffi-dev libssl-dev build-essential xvfb libboost-python-dev libleveldb1 libleveldb-dev libjpeg-dev
 if [ "$flash" = true ]; then
-    sudo apt-get install adobe-flashplugin
+    sudo apt-get install -y adobe-flashplugin
 fi
 
-wget https://bootstrap.pypa.io/get-pip.py
-sudo -H python get-pip.py
-rm get-pip.py
-
-sudo -H pip install -U setuptools
-sudo -H pip install -U pyvirtualdisplay beautifulsoup4 pyasn1 PyOPenSSL python-dateutil tld pyamf psutil pyhash plyvel tblib tabulate pytest publicsuffix
-
-# Install specific mitmproxy version since we rely on some internal structure of
-# netlib and mitmproxy. New releases tend to break things and should be tested
-sudo -H  pip install mitmproxy==0.13
-
-# Install specific version of selenium known to work well with the Firefox install below
-sudo -H pip install selenium==2.53.0
+# Check if we're running on continuous integration
+# Python requirements are already installed by .travis.yml on Travis
+if [ "$TRAVIS" != "true" ]; then
+	wget https://bootstrap.pypa.io/get-pip.py
+	sudo -H python get-pip.py
+	rm get-pip.py
+	sudo pip install -r requirements.txt
+fi
 
 # Install specific version of Firefox known to work well with the selenium version above
 wget https://ftp.mozilla.org/pub/firefox/releases/45.0.1/linux-x86_64/en-US/firefox-45.0.1.tar.bz2
